Add binary serializer for cheezy path
This is not fully tested for accuracy yet and I would not recommend switching the production code to use this serializer.

Benchmarking on my laptop vs the current plaintext serializer, the binary serializer finishes in 3ms while the text serializer finishes in 130ms.

The 130ms is a significant chunk of the total time to generate a path and serialize it. Not sure how significant it will be for the rio, though.

The deserializer is faster too.